bugfix/ErrorHandling & dividing generateAI file
