+++ src/app/(app)/dashboard/history/[id]/_actions/update-invoice-action.js
@@ -1,11 +1,11 @@
 "use server";
 
 import { prisma } from "@/utils/prisma";
-import { openai } from "@/utils/openai";
 import { redirect } from "next/navigation";
 import { resend } from "@/utils/resend";
 import path from "path";
 import fs from "fs";
+import { generateAI } from "@/libs/generateAI";
 
 export async function updateInvoiceAction(_, formData) {
   const name = formData.name;
@@ -27,63 +27,7 @@ export async function updateInvoiceAction(_, formData) {
   let newInvoice;
 
   if (type === "generate") {
-    const completion = await openai.chat.completions.create({
-      model: "deepseek/deepseek-chat:free",
-      messages: [
-        {
-          role: "system",
-          content: `You are the secretary/personal assistant of this user:
-          User:
-          name: ${userName}
-          organization/company: ${organization}
-          Email: ${senderEmail}
-
-          Currently, you are ordered by user to make a ${invoiceType} invoice cover letter to their client, with the following information:
-
-          Recipient:
-          organization/company: ${recipient}
-          Address: ${recipientAddress}
-          phone number: ${recipientPhone}
-          recipient email: ${recipientEmail}
-          item: ${invoiceItems.map((item) => {
-            return `${item.name} - item price :${item.price}`;
-          })}
-          quantity: ${invoiceItems.map((item) => {
-            return `${item.quantity}`;
-          })}
-          total prices: ${totalPrice},
-          bill due date: ${dueDate}
-
-          The results should be ready to use and will be sent to the client email
-
-          IMPORTANT: JUST A LETTER, NO EXPLANATIONS AND KEY IMPROVEMENTS, NO USER PHONE NUMBER, ALL CURRENCY IN INDONESIAN RUPIAH, AND RESPONSE ARE IN ENGLISH
-
-          OUPUT MUST BE IN VALID JSON FORMAT AND READY TO USE WITHOUT EDIT, WITHOUT ANY BACKTICKS (\`) AND WITHOUT '''JSON
-
-          Example:
-            {"subject":"email subject", "letter": "content you make(cover letter)"}
-          `,
-        },
-
-        {
-          role: "user",
-          content: `I want to give an invoice to my client, please make a cover letter for the subject and body of the email, i want an ${invoiceType} cover letter.`,
-        },
-      ],
-      temperature: 0.7,
-    });
-    const result = completion.choices[0].message.content;
-
-    try {
-      const parsedResult = JSON.parse(result);
-
-      return parsedResult;
-    } catch (error) {
-      console.log({
-        message: "Error parsing JSON",
-        error: error,
-      });
-    }
+    return await generateAI(formData);
   }
   if (type === "submit") {
     const subject = formData.subject;
@@ -161,7 +105,6 @@ export async function updateInvoiceAction(_, formData) {
         error: error,
       });
     }
+    redirect(`/dashboard/history/${newInvoice.id}`);
   }
-
-  redirect(`/dashboard/history/${newInvoice.id}`);
 }
